# Various minor performance improvements