--- a/src/seq2d.nim
+++ b/src/seq2d.nim
@@ -12,13 +12,13 @@ proc newSeq2D*[T](width, height: int): Seq2D[T] =
   result.width = width
   result.height = height
 
-proc `width`*[T](this: Seq2D[T]): int =
-  return this.width
+template `width`*[T](this: Seq2D[T]): int =
+  this.width
 
-proc `height`*[T](this: Seq2D[T]): int =
-  return this.height
+template `height`*[T](this: Seq2D[T]): int =
+  this.height
 
-proc `[]`*[T](this: Seq2D[T], x, y: int): T =
+proc `[]`*[T](this: Seq2D[T], x, y: int): T {.inline.} =
   when not defined(danger):
     if x >= this.width:
       raise newException(Exception, fmt"x value of {x} outside bounds")
@@ -26,7 +26,7 @@ proc `[]`*[T](this: Seq2D[T], x, y: int): T =
       raise newException(Exception, fmt"y value of {y} outside bounds")
   return this.data[x + this.width * y]
 
-proc `[]=`*[T](this: var Seq2D[T], x, y: int, t: T) =
+proc `[]=`*[T](this: var Seq2D[T], x, y: int, t: T) {.inline.} =
   when not defined(danger):
     if x >= this.width:
       raise newException(Exception, fmt"x value of {x} outside bounds")