As per the Nim forum topic I tried to optimise this a little bit. Didn't look at the algorithm at all, just applied some typical optimisation strategies (better compile flags, adding a main, inlining things, avoiding ref objects).

Times on my machine went from (range is min/max of three runs):
Generation: 6.2 - 6.7 ms down to 3.4 - 3.5 ms
Cold run: 9.9 - 10.4 ms down to 9.0 - 9.3 ms
Warm run: 9.9 - 10.1 ms down to 8.7 9.0 ms

So about a 10% speedup, 50% in the case of the generation. I also tried to preallocate the path stack with `newSeqOfCap[Point](3679)` but that didn't seem to give any noticeable speedup.